Renamed Files, code improved
