+++ .gitignore
@@ -1,12 +1,8 @@
-# OS generated files #
-######################
 .DS_Store
 .DS_Store?
 ._*
 .Trashes
 ehthumbs.db
 Thumbs.db
-
 *.pyc
-/Python Files/login.py
