+++ Python Files/login.py
@@ -1,0 +1,10 @@
+def get_credentials():
+	credentials = {
+		'username' : 'sbaveja',
+		'password' : '',
+		'dbname' : 'devdb12' 
+	}
+	return credentials
+
+if __name__ == "__main__":
+	get_credentials()
